requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=8.0
pytest-django>=4.8

# Pre-commit and linting tools
pre-commit>=3.5.0
ruff>=0.1.9
# Note: dotenv-linter is installed separately as a binary, not via pip